+++ config/babel.conf.wds.js
@@ -1,0 +1,13 @@
+const basicConfig = require('./babel.conf')
+
+module.exports = Object.assign({}, basicConfig, {
+  presets: [
+    ['@babel/env', {
+      targets: '> 0.25%, not dead',
+    }],
+    ...basicConfig.presets,
+  ],
+  plugins: [
+    ...basicConfig.plugins,
+    'react-hot-loader/babel'],
+})