+++ changelog.md
@@ -1,3 +1,10 @@
+## [0.10.1] - 11 Sep 2018
+### Changed
+  - separate babel configs for client and webpack
+### Fixed
+  - actual use of babel cofig by webpack-dev-server
+  - proper title in wds template
+
 ## [0.10.0] - 10 Sep 2018
 ### Added
   - configuring options. [Read more here](readme.md#Configuring)